Shifted BitVector to use size_t

+++ tools/BitVector.h
@@ -33,7 +33,7 @@ namespace emp {
     using field_t = uint64_t;
 #endif
 
-    static constexpr int FIELD_BITS = sizeof(field_t)*8;
+    static constexpr size_t FIELD_BITS = sizeof(field_t)*8;
     size_t num_bits;
     field_t * bit_set;
 
@@ -360,9 +360,9 @@ namespace emp {
 
 
     // Count 1's by looping through once for each bit equal to 1
-    int CountOnes_Sparse() const {
+    size_t CountOnes_Sparse() const {
       const size_t NUM_FIELDS = NumFields();
-      int bit_count = 0;
+      size_t bit_count = 0;
       for (size_t i = 0; i < NUM_FIELDS; i++) {
         field_t cur_field = bit_set[i];
         while (cur_field) {
@@ -374,10 +374,10 @@ namespace emp {
     }
 
     // Count 1's in semi-parallel; fastest for even 0's & 1's
-    int CountOnes_Mixed() const {
+    size_t CountOnes_Mixed() const {
       const size_t NUM_FIELDS = NumFields() * sizeof(field_t)/4;
       uint32_t* uint_bit_set = (uint32_t *) bit_set;
-      int bit_count = 0;
+      size_t bit_count = 0;
       for (size_t i = 0; i < NUM_FIELDS; i++) {
         const uint32_t v = uint_bit_set[i];
         const uint32_t t1 = v - ((v >> 1) & 0x55555555);
@@ -387,7 +387,7 @@ namespace emp {
       return bit_count;
     }
 
-    int CountOnes() const { return CountOnes_Mixed(); }
+    size_t CountOnes() const { return CountOnes_Mixed(); }
 
     int FindBit() const {
       const size_t NUM_FIELDS = NumFields();
